refactor: controller 분리되어 있던 부분 합침
